requirements.txt
streamlit >= 1.41.1
python-statemachine >= 2.5.0